no more shared initialization of nordic boards
