+++ boards/nordic/nrf52_components/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "nrf52dk_base"
+name = "nrf52_components"
 version = "0.1.0"
 authors = ["Tock Project Developers <tock-dev@googlegroups.com>"]
 edition = "2018"
