+++ boards/nordic/nrf52_components/src/lib.rs
@@ -1,0 +1,9 @@
+#![no_std]
+
+pub mod ble;
+pub mod startup;
+
+pub use self::ble::BLEComponent;
+pub use self::startup::{
+    NrfClockComponent, NrfStartupComponent, UartChannel, UartChannelComponent, UartPins,
+};
