+++ boards/nordic/nrf52dk/build.rs
@@ -1,5 +1,4 @@
 fn main() {
     println!("cargo:rerun-if-changed=layout.ld");
-    println!("cargo:rerun-if-changed=chip_layout.ld");
     println!("cargo:rerun-if-changed=../../kernel_layout.ld");
 }
