--- boards/nordic/nrf52dk_base/README.md
@@ -1,5 +1,0 @@
-Generic nRF52 Board
-===================================
-
-This crate has general code that is used by both the nRF52dk and nRF52840dk
-development boards.
