--- boards/nordic/nrf52dk_base/src/nrf52_components/mod.rs
@@ -1,5 +1,0 @@
-pub mod ble;
-pub mod startup;
-
-pub use self::ble::BLEComponent;
-pub use self::startup::{NrfClockComponent, NrfStartupComponent};
