+++ chips/nrf52832/src/lib.rs
@@ -1,7 +1,7 @@
 #![no_std]
 
 pub use nrf52::{
-    adc, aes, ble_radio, clock, constants, crt1, ficr, i2c, ieee802154_radio, init, nvmc,
+    acomp, adc, aes, ble_radio, clock, constants, crt1, ficr, i2c, ieee802154_radio, init, nvmc,
     peripheral_interrupts, pinmux, ppi, pwm, rtc, spi, temperature, timer, trng, uart, uicr,
 };
 pub mod chip;
